make test code simple
